improve solution so new props dont need to be passed to dropdown component
