+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-content/user-participants/component.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { TransitionGroup, CSSTransition } from 'react-transition-group';
-import PropTypes from 'prop-types';
 import { defineMessages } from 'react-intl';
+import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { styles } from '/imports/ui/components/user-list/user-list-content/styles';
 import UserListItem from './user-list-item/component';
@@ -47,38 +47,6 @@ const intlMessages = defineMessages({
   usersTitle: {
     id: 'app.userList.usersTitle',
     description: 'Title for the Header',
-  },
-  ChatLabel: {
-    id: 'app.userList.menu.chat.label',
-    description: 'Save the changes and close the settings menu',
-  },
-  ClearStatusLabel: {
-    id: 'app.userList.menu.clearStatus.label',
-    description: 'Clear the emoji status of this user',
-  },
-  MakePresenterLabel: {
-    id: 'app.userList.menu.makePresenter.label',
-    description: 'Set this user to be the presenter in this meeting',
-  },
-  RemoveUserLabel: {
-    id: 'app.userList.menu.removeUser.label',
-    description: 'Forcefully remove this user from the meeting',
-  },
-  MuteUserAudioLabel: {
-    id: 'app.userList.menu.muteUserAudio.label',
-    description: 'Forcefully mute this user',
-  },
-  UnmuteUserAudioLabel: {
-    id: 'app.userList.menu.unmuteUserAudio.label',
-    description: 'Forcefully unmute this user',
-  },
-  PromoteUserLabel: {
-    id: 'app.userList.menu.promoteUser.label',
-    description: 'Forcefully promote this viewer to a moderator',
-  },
-  DemoteUserLabel: {
-    id: 'app.userList.menu.demoteUser.label',
-    description: 'Forcefully demote this moderator to a viewer',
   },
 });
 
@@ -136,7 +104,6 @@ class UserParticipants extends Component {
       normalizeEmojiName,
       isMeetingLocked,
       users,
-      intl,
       changeRole,
       assignPresenter,
       setEmojiStatus,
@@ -146,50 +113,6 @@ class UserParticipants extends Component {
       getEmojiList,
       getEmoji,
     } = this.props;
-
-    const userActions =
-    {
-      openChat: {
-        label: () => intl.formatMessage(intlMessages.ChatLabel),
-        handler: (router, user) => router.push(`/users/chat/${user.id}`),
-        icon: 'chat',
-      },
-      clearStatus: {
-        label: () => intl.formatMessage(intlMessages.ClearStatusLabel),
-        handler: user => setEmojiStatus(user.id, 'none'),
-        icon: 'clear_status',
-      },
-      setPresenter: {
-        label: () => intl.formatMessage(intlMessages.MakePresenterLabel),
-        handler: user => assignPresenter(user.id),
-        icon: 'presentation',
-      },
-      remove: {
-        label: user => intl.formatMessage(intlMessages.RemoveUserLabel, { 0: user.name }),
-        handler: user => removeUser(user.id),
-        icon: 'circle_close',
-      },
-      mute: {
-        label: () => intl.formatMessage(intlMessages.MuteUserAudioLabel),
-        handler: user => toggleVoice(user.id),
-        icon: 'mute',
-      },
-      unmute: {
-        label: () => intl.formatMessage(intlMessages.UnmuteUserAudioLabel),
-        handler: user => toggleVoice(user.id),
-        icon: 'unmute',
-      },
-      promote: {
-        label: () => intl.formatMessage(intlMessages.PromoteUserLabel),
-        handler: user => changeRole(user.id, 'MODERATOR'),
-        icon: 'promote',
-      },
-      demote: {
-        label: () => intl.formatMessage(intlMessages.DemoteUserLabel),
-        handler: user => changeRole(user.id, 'VIEWER'),
-        icon: 'user',
-      },
-    };
 
     let index = -1;
 
@@ -210,7 +133,6 @@ class UserParticipants extends Component {
             isBreakoutRoom={isBreakoutRoom}
             user={user}
             currentUser={currentUser}
-            userActions={userActions}
             meeting={meeting}
             getAvailableActions={getAvailableActions}
             normalizeEmojiName={normalizeEmojiName}
@@ -219,6 +141,11 @@ class UserParticipants extends Component {
             handleEmojiChange={handleEmojiChange}
             getEmojiList={getEmojiList}
             getEmoji={getEmoji}
+            setEmojiStatus={setEmojiStatus}
+            assignPresenter={assignPresenter}
+            removeUser={removeUser}
+            toggleVoice={toggleVoice}
+            changeRole={changeRole}
           />
         </div>
       </CSSTransition>
