--- bigbluebutton-html5/imports/ui/components/user-list/user-list-content/user-participants/user-list-item/styles.scss
@@ -1,7 +1,0 @@
-@import "/imports/ui/stylesheets/variables/_all";
-
-.emojiSelected {
-    span, i {
-      color: $color-primary;
-    }
-  }
