+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-content/user-participants/user-list-item/user-dropdown/styles.scss
@@ -103,3 +103,9 @@
   max-width: 100%;
   overflow: visible;
 }
+
+.emojiSelected {
+  span, i {
+    color: $color-primary;
+  }
+}
